// src/web/src/api/generated/model/salesChannelDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Zalihe.Web | v1
 * OpenAPI spec version: 1.0.0
 */
import type { SalesChannelStatus } from './salesChannelStatus';
import type { SalesChannelType } from './salesChannelType';

export interface SalesChannelDto {
  id: string;
  type: SalesChannelType;
  baseUrl: string;
  status: SalesChannelStatus;
  /** @nullable */
  lastErrorCode: string | null;
  /** @nullable */
  lastSyncedAt: string | null;
  createdAt: string;
}